requirements.txt
fastapi==0.76.0
keras==2.8.0
matplotlib==3.5.1
numpy==1.22.2
Pillow==9.1.0
tensorflow==2.8.0
